Altering the Idol of Kozah swaps
The previous replace would match !PartyhasItem("misc48" with the first match, resulting in triggers that would end up as
!OR(2) PartyHasItem("MISC48") PartyHasItem("C0AUIM04"). I've added a \b check to avoid this.

I've also moved the replaces to a dialogue file; this reduces the scope of the search  to just dialogue triggers instead of
searching all of the text.

Added an UNLESS trigger so that it will skip dialogues if they already account for the alternative item file. E.g. Glam's
NPC Pack was getting errors because the patch would result in nested ORs so it now accounts for the alternative item
directly and does not need to be patched.

https://www.gibberlings3.net/forums/topic/26699-report-bugs-and-typos-here/page/5/#findComment-306884